Bug 232063 - Relationships between JPA 2.0 Cachable(false) and Cachable(true) Entities should be supported Checkin 2 Resolves issues with MissingDescriptorListener

+++ jpa/eclipselink.jpa.test/src/org/eclipse/persistence/testing/tests/jpa/xml/advanced/EntityMappingsAdvancedJUnitTestCase.java
@@ -831,7 +831,7 @@ public class EntityMappingsAdvancedJUnitTestCase extends JUnitTestCase {
 
     public void testReadEmployee() {
         Employee employee = createEntityManager(m_persistenceUnit).find(Employee.class, employeeId);
-        assertTrue("Error reading Employee", employee.getId() == employeeId);
+        assertTrue("Error reading Employee", employee.getId() == employeeId.intValue());
     }
 
     public void testUpdateEmployee() {
